fix: Changed sound check to use isRift method refactor: Moved isRift into RiftLevelManager
