+++ src/main/java/com/wanderersoftherift/wotr/fix/DisableTrialSpawnerAmbient.java
@@ -1,6 +1,7 @@
 package com.wanderersoftherift.wotr.fix;
 
 import com.wanderersoftherift.wotr.WanderersOfTheRift;
+import com.wanderersoftherift.wotr.core.rift.RiftLevelManager;
 import net.minecraft.client.Minecraft;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.sounds.SoundEvents;
@@ -31,7 +32,7 @@ public final class DisableTrialSpawnerAmbient {
     @SubscribeEvent
     public static void catchPlaySound(PlaySoundEvent event) {
         if (TRIAL_AMBIENT_SOUNDS.contains(event.getOriginalSound().getLocation())
-                && Minecraft.getInstance().level.dimension().location().getPath().startsWith("rift_")) {
+                && RiftLevelManager.isRift(Minecraft.getInstance().level)) {
             event.setSound(null);
         }
     }
