+++ src/main/java/com/wanderersoftherift/wotr/loot/LootUtil.java
@@ -1,6 +1,7 @@
 package com.wanderersoftherift.wotr.loot;
 
 import com.wanderersoftherift.wotr.core.rift.RiftData;
+import com.wanderersoftherift.wotr.core.rift.RiftLevelManager;
 import com.wanderersoftherift.wotr.init.ModLootContextParams;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.level.storage.loot.LootContext;
@@ -11,7 +12,7 @@ public class LootUtil {
         Integer riftTier = context.getOptionalParameter(ModLootContextParams.RIFT_TIER);
         if (riftTier == null) {
             ServerLevel serverlevel = context.getLevel();
-            if (!RiftData.isRift(serverlevel)) {
+            if (!RiftLevelManager.isRift(serverlevel)) {
                 return 0;
             }
             riftTier = RiftData.get(serverlevel).getTier();
