+++ src/main/java/com/wanderersoftherift/wotr/loot/lootmodifiers/OverrideVanillaLootModifier.java
@@ -2,7 +2,7 @@ package com.wanderersoftherift.wotr.loot.lootmodifiers;
 
 import com.mojang.serialization.MapCodec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
-import com.wanderersoftherift.wotr.core.rift.RiftData;
+import com.wanderersoftherift.wotr.core.rift.RiftLevelManager;
 import com.wanderersoftherift.wotr.init.ModDataComponentType;
 import com.wanderersoftherift.wotr.item.implicit.GearImplicits;
 import com.wanderersoftherift.wotr.item.socket.GearSockets;
@@ -37,7 +37,7 @@ public class OverrideVanillaLootModifier extends LootModifier {
         // Check if the current level is a rift level and return
         // note this should never be the case, as the conditions from glm should prevent it
         ServerLevel serverlevel = context.getLevel();
-        if (RiftData.isRift(serverlevel)) {
+        if (RiftLevelManager.isRift(serverlevel)) {
             return generatedLoot;
         }
         ResourceKey<Level> dimension = context.getLevel().dimension();
